+++ misc/python/materialize/cloudtest/app/materialize_application.py
@@ -16,7 +16,7 @@ from typing import Optional
 from pg8000.exceptions import InterfaceError
 
 from materialize import ROOT, mzbuild
-from materialize.cloudtest.application import Application
+from materialize.cloudtest.app.application import Application
 from materialize.cloudtest.k8s.cockroach import COCKROACH_RESOURCES
 from materialize.cloudtest.k8s.debezium import DEBEZIUM_RESOURCES
 from materialize.cloudtest.k8s.environmentd import (
