+++ test/cloudtest/conftest.py
@@ -14,7 +14,7 @@ if TYPE_CHECKING:
 
 import pytest
 
-from materialize.cloudtest.materialize_application import MaterializeApplication
+from materialize.cloudtest.app.materialize_application import MaterializeApplication
 
 
 def pytest_configure(config: "Config") -> None:
