+++ test/cloudtest/test_antiaffinity.py
@@ -9,8 +9,8 @@
 
 import pytest
 
+from materialize.cloudtest.app.materialize_application import MaterializeApplication
 from materialize.cloudtest.k8s import cluster_pod_name
-from materialize.cloudtest.materialize_application import MaterializeApplication
 from materialize.cloudtest.wait import wait
