+++ test/cloudtest/test_compute.py
@@ -10,9 +10,9 @@
 import pytest
 from pg8000.exceptions import InterfaceError
 
+from materialize.cloudtest.app.materialize_application import MaterializeApplication
 from materialize.cloudtest.exists import exists, not_exists
 from materialize.cloudtest.k8s import cluster_pod_name, cluster_service_name
-from materialize.cloudtest.materialize_application import MaterializeApplication
 from materialize.cloudtest.wait import wait
